reader-map: evict number of keys given to RandomEviction
Updates RandomEviction::pick_keys_to_evict() and
RandomEviction::pick_ranges_to_evict() such that exactly the number of
keys requested are chosen.

Previously, given k keys to evict and N total keys in the map,
RandomEviction would evict each key in the map independently with
probability k/N. This strategy could evict more or less keys than the
number requested.

RandomEviction now allocates an array of indices and randomly shuffles
the set, evicting any key with a shuffled index that is less than the
number of requested keys to evict. This strategy guarantees that exactly
the number of requested keys are evicted.

The new method for selecting keys is more expensive than the previous
method, requiring an allocation equal to the length N of the total keys
in the map and a shuffle which requires N random access operations on
the index vector.

There is unfortunately no practical way to create an iterator yielding a
random permutation of another iterator, without creating a new
allocation.

Change-Id: Idd3febad343904830a9299f87e6891a4e2c71ab5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5138
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>