benchmarks/write-benchmark: Make offered load fully open
Previously, the write benchmark ran a loop per thread that:

  1. slept up to the next (1/qps) interval
  2. performed a write, measuring the latency

The way this worked, if a single write takes longer than 1/qps we
would simply throttle down to a maximum qps of 1/latency, preventing
measuring the impact of true saturation of readyset's ability to handle
writes.

This commit changes the behavior of the write benchmark to instead spawn
a tokio task for each write every 1/qps, using the new connection pool
support in database-utils to give each of those tasks its own
connection. Ideally this should give a more accurate sense of the impact
on latency of actually offering more and more write load to ReadySet.

Change-Id: Icaf352f646a4f79ecd948056a4cd1d211a6cf02a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5142
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>