reader-map: Return the evicted key for single key evictions
Adds a SingleKey variant to EvictionRequest, which is intended to be
used in tests to evict a single key. When the reader evicts a single key
from state per this request, it will now return the key that was
evicted. This key is propagated back up to Domain::handle_eviction(),
but not yet used.

The conditional returning of the evicted key based on the request is not
particularly neat. However, this is done to avoid always returning the
key which would require an unecessary clone, and to avoid creating a
separate code path for that particular request.

Change-Id: I8563e638aeb4947200b46be082f8d5f1441136eb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5163
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>
Reviewed-by: Nick Marino <nick@readyset.io>