Controller: return evicted key from /evict_single rpc
Renames /evict_random to /evict_single. Creates a SingleKeyEviction
struct containing the Key, Tag, and DomainIndex used in an eviction.
This struct is returned from a call to the /evict_single rpc if an
eviction occurred.

The DomainIndex is included to avoid looking up the Domain given a Tag,
since we don't store a direct mapping in the controller.

The Tag is represented as a u32 to avoid the cyclic dependency between
readyset-client and readyset-common.

Change-Id: I9ffa4be516f95454d0df220ad65268aae1a0a33a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5165
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>