clustertest: Fix duration panic on time out
While waiting for an authority during clustertests, if the timeout is
exceeded after restarting the loop, but before actually blocking
in tokio, we might see an elapsed duration which is larger than the
timeout. The inline code that currently checks this does a simple
subtraction of Durations. Internal to the Duration Sub impl, is a check
that the result is not negative, and will panic if it is negative.

We already had a similar check for this in
utils::query_until_expected_inner(), but it could still hit the same
Duration subtraction problem (although less likely). Both code paths
are cleaned up and safer wrt subtracting Durations.

Change-Id: I49e783682a989f858c5e6a482bfb5f9960a48911
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5158
Reviewed-by: Luke Osborne <luke.o@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI