DfState: allow external evictions by key
Updates DfState::evict_single() to accept an optional SingleKeyEviction
argument. This will allow external requests for evictions by key.
Previously, we only triggered evictions by key within a Domain as a
result of an upstream eviction.

evict_single() accepts a Vec<DfValue> key which it must convert to a
KeyComparison when delegating to the underlying state. It is necessary
to create a KeyComparison::Range when delegating to a BTree reader map,
or else the map will not update its intervals. This is not the case for
a MemoryState with an underlying BTree map.

Change-Id: I32680fa3a673d8199359ad6117a62befa015b793
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5166
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>