vertical tests: support reproducible evictions for shrinking
Adds an inner: RefCell<Option<SingleKeyEviction>> field to
Operation::Evict. This field stores the body of the /evict_single rpc.

The rpc is first called with a body of None, which triggers an eviction
for a random key in a random partially materialized state. That rpc call
returns a SingleKeyEviction (if an eviction occurred) which is used to
update the inner value of that Operation::Evict.

Storing the SingleKeyEviction allows us to exactly reproduce a test run,
which is necessary for shrinking a failing test case to the set of
minimal reproducible steps.

If the SingleKeyEviction targets a key that has been removed by
shrinking, the dataflow state will be unchanged and the test will
continue.

Change-Id: I93be3b95ce161cbb7cf889b240bcc06935ff360b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5168
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>