readyset-server: Bail if postgres upstream url lacks a database name
Require that a postgres upstream URL contain a database name, e.g.:
postgres://user:pass@host:port/database_name
                               ^------------

Typical postgres client behavior is to assume a database name equal to
the username if none is provided. However, we are diverging from this
behavior because we consider it a footgun.

Fixes: REA-2917
Change-Id: I4561c21d1d189dcf247881e6d20d1af1823d70ca
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5197
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>