system-benchmarks: Include database name
Postgres requires that we always include a database name when initiating
a connection, and a constraint was added to `DatabaseURL` to enforce
this. There were a few places where we were parsing a `DatabaseURL` from
a string without a database name, so this commit adds database names in
those cases.

Change-Id: Ibbe5e036978aa59aec432e338981da74096a7f16
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5205
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>