Remove metrics-aggregator and supporting code
The metrics-aggregator application has been unused since its original
introduction, and especially since we now have `SHOW CACHES` and `SHOW
PROXIED QUERIES` it isn't really carrying its weight, especially as a
distinct service that has to be run. This commit removes the
metrics-aggregator crate itself, along with the code in the authority
for registering adapter endpoints which only existed to support it, and
the `allow-list` and `deny-list` endpoints in the adapter which only
existed to be queried by it.

Change-Id: I12b1ab58b7c3bd351f6dfa0f3ca8434bb23facea
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5209
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke.o@readyset.io>