nom-sql: impl Arbitrary for Expr
Implement proptest::Arbitrary for nom_sql::Expr, doing so with a
manual rather than a derived impl so that generating exprs doesn't stack
overflow. That comes with a *very* simple proptest which simply
generates an arbitrary Expr but does nothing with it, to check that the
generation process doesn't ever stack overflow.

Change-Id: Ib1c2a55aaf035af6fb2bc868e7602abe7982e2c2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5267
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>