sql: Do the job of count_star_rewrite in qg->mir
Rather than doing the rewrite from `COUNT(*)` to
`COUNT(coalesce(some_col, 0))` (which we need to do since all
aggregates) in the AST as a rewrite pass, do essentially the same work
during qg->mir conversion, by picking an arbitrary column in the parent
node and adding a project node for the coalesce call before the
aggregate node itself. This is not just a fair bit less code, but also
fixes an issue where queries where there isn't a good column to pick (eg
ones where we only select unaliased expressions in subqueries; see the
logictest in the diff for an example) would cause a panic in the
previous AST count_star_rewrite pass.

Fixes: REA-3039
Change-Id: Ib31fac8d853e8af5dfd4c27c8589f314e630ca17
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5292
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>