dataflow-state: Vary bench reads by count
Previously, the persistent state benchmarks did not factor in the batch
size when choosing which reads to perform: if the number of entries in
the database was increased by a factor of 10, the benchmarks only would
have read from a subset of the key space. This commit updates the
benchmarks to factor in the size of the database when choosing what
reads to perform.

Change-Id: I64482a7a1f6041a936370d6c5ac6048d1ad593af
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5313
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>