dataflow-state: Insert in batches in benches
Previously, the persistent state benchmarks were inserting data into the
persistent state instance one row at a time. This was quite slow,
resulting in the benchmarks taking a very long time to run for larger
datasets. This commit changes our approach to instead insert data in
batches of 100,000 rows, vastly speeding up the amount of time it takes
to seed the database.

Change-Id: I316553f53af7b1c55fbf5bfbed7aaffb6ecd230b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5317
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>