nom-sql: Allow literal formatting to be dialect-driven
This commit does the same thing for nom_sql::Literal that we've done
for most of the other types in the AST where we have a `display(Dialect)
-> impl fmt::Display` method rather than actually impl'ing fmt::Display.
This is done to prepare for some dialect-specific formatting we're going
to need to do (specifically for hex literals). None of that actually
happens in this commit yet, though - this is just a refactor to prepare.

Change-Id: Iaa03e63e5748a20635244fc7ee96f12b80c18eff
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5322
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>