dataflow: don't propagate filter expr errors
If expression evaluation fails in the filter node, we now consider the
expression to evaluate to false. Previously we propagated these errors
and killed the domain.

Ensures that we log this error once per on_input() call if it occurs in
the filter or project node.

Expression evaluation can still fail in query_through(), because this
will not cause the domain to fail.

Ref: REA-2964

Change-Id: I38d56f23049309fb2dac464dc6a1169a8d354ba6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5336
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>