query_graph: Collect extra preds in join conditions
This commit slightly relaxes the limitation on join conditions where all
conditions must be representable as an equi-join-key. While handling
join predicates, we collect predicates that aren't equal comparisons on
columns into an `extra_preds` vector. For INNER joins, we can handle
those the same as the predicates in the WHERE clause (by classifying
them as either local predicates, global predicates, or parameters). For
LEFT joins, those predicates will need to be handled specially - for
now, we store them as a field on `QueryGraphEdge::LeftJoin` and return
an unsupported error during qg->mir lowerinng if that field is
non-empty.

Fixes: REA-3070
Change-Id: Ib793affbea805b3761d4469acb3427f07325bcfd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5366
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>