proptest-stateful: Add macro for using better failure persistence defaults
This is a convenience macro that is useful for creating ProptestConfig
structs with failure persistence fields that match the current source
file. This is useful for stateful property tests because leaving these
fields at their default values causes them to be set by the proptest
macros invoked in proptest_stateful::test, which is not typically what
one would actually want.

Change-Id: I02c4327b3cc955ca24e4dfb0c5750353e71eb0a8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5373
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>