ddl_vertical: Fix precondition bug for WriteRow
Without this fix in place, we can sometimes incorrectly shrink out steps
that add/remove columns from the end of a table, because the
precondition uses `zip` to compare the types in the table definition and
the WriteRow op, and `zip` will silently drop any elements in the tail
of the longer iterator that fall past the end of the shorter iterator.

This would then result in failures due to things like inserting the
wrong number of columns into a table, because the test wouldn't notice
that we can't shrink out an AddColumn and still write a value to that
column in a later WriteRow.

Change-Id: Ib58a873090ee18ebc6e4ac3195955a564289168f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5376
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>