replicators: Wait for DB to not be in use
As faar as I can tell, it looks like some of the time the replicator
thread that's part of one replicator test doesn't fully shut down before
another starts, meaning pg thinks the database that we're testing
against is still in use. Without fully diagnosing why that's happening,
this commit just makes us re-try the `DROP DATABASE IF EXISTS` every
100ms until it stops returning that error. This should fix all of the
sources of test flakiness we've seen in this test suite in nightly.

Fixes: REA-2878
Fixes: REA-3061
Change-Id: I41ad9a49e716ef0e002620aa2ed49b281b83b9d4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5382
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI