Adapter: Allow dropping caches in implicit modes
If DROP CACHE is run in an implicit caching mode, the dropped query will
now be deny-listed until it is explicitly cached again. Previously, the
query would be implicitly cached again the next time it was run.

This is implemented with a new MigrationState::Dropped variant.  For the
purpose of reporting query support to the user, dropped caches are
considered to be supported. For this reason, it is important to only set
a cached queries' state to dropped. This is currently loosely enforced
in the adapter by only updating a query's status to dropped after a
corresponding view for the query has been removed.

We could strictly enforce that a query can only transition to a dropped
state from a successful state; however, this logic may not work in the
case that the query was cached from another adapter, or the adapter it
was cached from has just restarted without seeing that query again.

Clearing the query status cache will not set all successful queries'
statuses to dropped, since `DROP ALL CACHES` is not obviously intended
to deny-list all cached queries.

Change-Id: Ia1a6bc837f9f314fadf85d927b060aee851e0c22
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5393
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI