ddl_vertical: Fix non-deterministic test generation of enum values
I was surprised to find that the Rust HashSet implementation doesn't
always result in the same order of elements when you convert it to a
Vec, even if you use the exact same input to create the same HashSet
twice; something internal to the HashSet seems to have some
non-determinism when it comes to the way it's represented in memory or
something, which then results in unpredictable ordering of elements when
you iterate over it.

The trouble is, this means that the same test seed can end up generating
different test cases. To fix this, we sort the result of
`gen_enum_values` and then call `prop_shuffle` on it. This we we still
get a randomly ordered set of enum values, but the same seed will
deterministically result in the same ordering of values.

Change-Id: Id561022a2a8ae6dd58b85b46617d538bf8b90d79
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5398
Reviewed-by: Aspen Smith <aspen@readyset.io>
Tested-by: Buildkite CI