ddl_vertical: Use BTreeMap instead of HashMap for determinism
There are a number of places in the test suite that we use `keys()` to
iterate over the keys of a map, and it turns out the order can randomly
vary from run to run, even if the inputs used to create a HashMap are
exactly the same.

This causes problems when we want to reproduce a test failure from a
given seed.

To fix this, I'm simply replacing every use of HashMap with BTreeMap,
since the API is essentially the same and BTreeMap gives us a reliably
deterministic ordering of keys.

Change-Id: I2c5a30d477bdec85823bc4e24c0bf5357524db09
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5399
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>