ddl_vertical: Fix shrinking bug around enum value insertions
Prior to this fix, it's possible we may incorrectly shrink a case like
this:

CREATE TYPE e AS ENUM ('a');
ALTER TYPE e ADD VALUE 'b';
CREATE TABLE t(enum_col e);
INSERT INTO t VALUES ('b');

The issue is that we may remove the ALTER TYPE but then still try to
insert 'b' into 't', resulting in an error since 'b' would no longer be
a valid enum value if we shrink out the second test step.

To fix this, I'm extending the preconditions for WriteRow to verify that
any enum values being written to a table match the current enum
definition of the corresponding type.

Change-Id: I4709553a15ac223fc98ec02247c6c8080b82a1a3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5404
Tested-by: Buildkite CI
Reviewed-by: Aspen Smith <aspen@readyset.io>