adapter: Serve SHOW CACHES via QueryStatusCache
Populate the results for the `SHOW CACHES` query from the adapter's
query status cache, rather than via the verbose_views controller RPC.

This makes it so the query texts match exactly between the results of
`SHOW PROXIED QUERIES` and `SHOW CACHES`, which is a user experience
win.

The drawback is that, on adapter restart, queries will not show up in
the `SHOW CACHES` results until the next time they are executed. We now
deem this an acceptable tradeoff.

Additionally, when dropping caches, the adapter only knows about caches
that it drops (and not e.g. those dropped via another adapter, or
automatically due to a replication error). This is because
ViewsSynchronizer doesn't currently poll for dropped caches. This had
implications for some existing unit tests.

Refs: REA-3055
Change-Id: Ibb9d59450254bf5511cc2623a0c3d994a282c09e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5408
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>