server: Allow some domains to be unplaced during recovery
If we ran recovery on a deployment when some domains were unschedulable,
eg because they had placement restrictions targeting workers that were
not yet in the cluster, previously we'd return an error that'd bubble
all the way up to the controller and crash the controller. Now, instead,
we allow some domains to be unscheduled during recovery and keep the
controller healthy. Some things needed to fall into place to make this
work:

1. We now track a list of domains that weren't placeable during recovery
   planning in a new field on `domain_migration_plan`
2. If that list is non-empty after recovery planning, we don't set
   `pending_recovery` on the Leader to `false` (note that this means we
   won't be able to handle RPCs that require quorum like
   `replication_offsets`, even though we ostensibly *could* - I've
   opened REA-3103 to fix that).
3. Since we'll now have multiple rounds of recovery planning (every time
   a worker joins until all domains are placed) we need to limit the set
   of domains we're trying to recover to only those which aren't already
   running on a worker.

This is all of course only done for *recovery* as of this commit, and
the original issue still exists for *migrations* (this can eg happen if
we're running with replicated readers and migrate a query while fewer
than the number of expected adapters has join the cluster). Fixing
that will come later.

Fixes: REA-3009
Change-Id: Idd101b7fad9b313c3910a3227663f138ffd5aedf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5410
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>