server: Don't require min_workers for any RPC
As part of making multi-node ReadySet deployments HA, we want to move
towards a more granular notion of the "health" of a deployment.
Specifically, it's desirable to make it so we can do as much work as
possible with the cluster (such as running the replicator!) even if not
all domains are running on workers.

To that end, this commit removes the `check_min_workers` macro from
`external_request` and merges all request matches into one big pattern
match, so that all controller RPCs can be done without having all
workers present or all domains running.

This is a *little* spooky, overall - I'm not sure what invariants might
be lurking in weird corners of the system that would cause this to
explode - but I've tested this at least with replicating changes (DDL or
otherwise) via the replicator when running with embedded readers and
without an adapter running, and things *seem* to work fine so far -
going forwards I'm going to be (both manually and via a clustertest)
testing a deployment with one adapter but configured to replicate
embedded readers twice, to further ensure this is all kosher.

Fixes: REA-3103
Change-Id: I3bf91be1132bc9b00c43edc801fd7a849877dbab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5412
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>
Tested-by: Buildkite CI