server: Make worker URL for a domain/shard/replica an Option
In preparation for making domain placement success granular on a per
shard+replica basis rather than all-or-nothing, make the URL of the
worker within DomainHandle (propagating out to a bunch of other fields)
an Option. This is not yet ever able to be None - that'll come next once
we teach the scheduler to set it to None if scheduling a shard+replica
fails.

Refs: REA-3106
Change-Id: I32832653257fe472a418d6f700f55c73e8559972
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5413
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>