client: Make addrs in ReaderHandleBuilder Options
To pave the way for passing around reader handle (builders) for readers
whose replicas are not all running, make the socket addrs in
ReaderHandleBuilder::replica_shard_addrs Options. When building an
actual reader handle for a replica, if any of the addrs for the shards
are None, we'll return a new ReadySetError variant indicating that we
can't talk to this replica of this reader (when we decide to bring back
sharding we likely want to make this *even more granular*, by allowing
talking to some shards of a replica when other shards aren't running,
but for now we're only running with 1 shard so this is fine).

Nothing currently ever sets this Option to None - that'll be coming up
next.

Change-Id: Id2bdac7d67c99d14118b738a836ad8b54c42ab37
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5420
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Luke Osborne <luke@readyset.io>