server: Base recovery needed off of all domains being placed
Rather than (functionally) denormalizing the information about whether
all domains in a deployment are running on workers to the
`pending_recovery` field on Leader, just calculate it when we need it
based on looking to see if all domains' replicas have been assigned to a
worker. This is ostensibly a little slower, but it's far from a hot path
and doing it this way allows keeping the state up to date much easier as
we make this more granular.

Change-Id: I50058c2ce36d2d29e1b12eb5cced996ff1656a61
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5421
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>