server: Allow requests to succeed for a subset of replicas
Allow scheduling domains, and then requests made *to* those domains, to
succeed for some shard-replicas of those domains but not others. To
enable this, `DomainHandle::send_to_healthy` now returns
Array2<Option<R>>` instead of `Array2<R>` (plus a similar change for
`send_to_healthy_shard` and `send_to_healthy_shard_replica`). For domain
requests such as `replication_offsets` that really *do* want to make
sure they get a successful, happy response from all replicas (and fail
if any replicas aren't placed onto a domain), there is now a
`DomainHandle::send_to_all` method as well.

Since scheduling can now happen for some shard-replicas of a domain but
not others, the overall success of planning a domain migration plan is
now at that level of granularity, both for recovery and for migration.

I've tested this manually by running a deployment with embedded readers
that's configured to replicate all readers twice, but that's only
running a single adapter instance - before this commit, migrating a new
query to that deployment would fail with an error, but now it works
perfectly fine but one of the replicas of the reader domain fails to
schedule as expected.

Since this is essentially removing the behavior that it was testing,
this commit removes the `assign_nonreader_domains_to_nonreader_workers`
test - the worker assignment behavior will be tested elsewhere later via
clustertests.

Fixes: REA-3106
Change-Id: I518059953d6e7b6285ffd7038d42a4e6ce1fc2bc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5422
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>