ddl_vertical: Fix test failures due to stale client type cache
Although the ReadySet failure was fixed in this CL:

https://gerrit.readyset.name/c/readyset/+/5431/3

The vertical DDL test still fails in that test case because the type
cache is also stale in the clients that the test is using.

Unlike in ReadySet though, it's easier to know when our client caches
might get stale, so we can just clear them as needed after type-altering
ops get run, rather than having to constantly assume they might need
clearing before every prepare.

Refs: REA-3108
Change-Id: I51fb8fe2a9d53d472543cef23f79aa8ab0efca1b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5450
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Aspen Smith <aspen@readyset.io>