server: Don't crash if we don't know the addr for a domain
When a new worker joins the cluster, we collect the addresses of all the
domains to send to that worker - previously, however, if we didn't yet
know the address of any domain, we'd return an error, which would bubble
up and crash the server. Now we just skip over that domain when
collecting addresses to send.

Change-Id: I1619e1859d1ec1642a0c38b02a411f471e0dfb3e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5453
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>