readyset-alloc: Add per-thread memory map
This ports over the thread memory map and related utilities from
tikv_util into readyset-alloc. They are separate in the tikv codebase
but it makes sense to bundle them here as we will be only using the
memory tracking functionality, which fits with the crate responsibility.

The changes ported over introduce a global static `THREAD_MEMORY_MAP`
that threads register to and deregister from when spawning automatically
if they use the `spawn_wrapper` instead of `spawn`. Updating our
existing `spawn`s for which we want to track memory to use
`spawn_wrapper` instead will be done separately.

Change-Id: I17ecbb41305a26af2663f6379a44b4ac0d08756d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5456
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>