clustertests: Remove postgres feature flag
It's unclear to me what the full original reasoning was for the feature
flag for postgres support in clustertests, but regardless it seems
unnecessary now. This commit removes that, and also the bit where we
load the database type from the env - with both of those changes, we can
just run both mysql and postgresql clustertests in a single run without
any conditional compilation, and save on double-compiling everything.

Change-Id: Ic715790666f414ec2ed0f5b3307c60a75f49c3a6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5152
Tested-by: Buildkite CI
Reviewed-by: Luke Osborne <luke@readyset.io>