readyset-adapter+server: Add memory stats http endpoints
This adds http endpoints to obtain stat dumps for memory usage for the
readyset adapter and server. The intention for these is to aid
debugging memory issues.

Note that the per-thread memory usage will be empty until the threads
are hooked up to the spawn_wrappers, which hasn't been done yet.

Change-Id: I16e0c54086d65a33c216b22efae55ec1618cc7dc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5462
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>