view: Don't pick non-scheduled replicas
If ReaderHandleBuilder::build is called without specifying a particular
replica, we'll just pick an arbitrary replica to use to build the
ReaderHandle. However, if some replicas are scheduled and others aren't,
this can result in us picking a replica without any addresses associated
with it and then failing later, even though another replica could easily
be built successfully. To fix that, this commit instead first filters
down to replicas who have all their shards scheduled before picking an
arbitrary one to build a ReaderHandle out of.

Change-Id: If5c11847aff983ef926b689fc355e8641b8a916e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5457
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>