clustertest: Add test for adapters < reader_replicas
Add a new clustertest to readyset_postgres for testing the case where
the number of scheduled adapters is less than the number of reader
replicas configured on the deployment. Currently this just checks that a
single fully materialized query can receive reads, but in the future
this can be extended to test more functionality.

Change-Id: I5fdd4e9873d37b794ff1a5d08633922cee398345
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5459
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>