Revert "readyset-psql: Decode enums as indices, then convert"
In order to avoid issues with enum types being altered and then decoded
incorrectly, we're switching back to decoding Postgres enums as their
labels at the protocol level, and only converting them to indices once
we get deeper into ReadySet.

Note that this doesn't solve the problem for arrays of enums, but that
will be addressed in a separate ticket and commit.

This reverts commit aced61125be47fc2963b8f3ba1ef3ab4faa47b75.

Fixes: REA-3108
Change-Id: I905ecf4cec895d08ef61dfc12df04cead94aa8a7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5464
Reviewed-by: Aspen Smith <aspen@readyset.io>
Tested-by: Buildkite CI