treewide: Register threads with memory tracker
This registers individually spawned threads with `spawn_wrapper` and
registers threadpools (currently only tokio runtimes) with
`with_sys_hooks` so that their memory usage is tracked by our
`THREAD_MEMORY_MAP`.

Synchornization only happens when threads (so the tokio runtime itself,
not individual tokio tasks) are spawned or dropped, or if a thread is
reading the memory map. Reading the memory map does not block registered
threads.

One limitation of this tracking is that threadpools that run
heterogeneous workloads (e.g. the adapter or server tokio runtimes, as
we are using them) do not have granular enough memory tracking to know
which tasks are using the most memory.

Release-Note-Core: Added a mechanism to insspect per-thread memory
usage.

Change-Id: I478def0379bed12bd50a5349d493fa932bd41261
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/5469
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>